Build using a local sdk directory
Trying to fix bugs is annoying when the problem is in the sdk because
you have to commit and push for each change. This patch adds a
"replace" to go.mod so the ../sdk/ directory is used. Because docker
can't see directories above the current one, we copy the sdk directory
to ./build-tmp each time we build.

Signed-off-by: Toby Cabot <toby@acnodal.io>