Improve Dockerfile caching
Evidently Docker's "ADD" command downloads the target to compute the
checksum to see if it can use a cached layer. This means that the file
is downloaded every build, which is wasteful, both of bandwidth and
time.

This patch replaces the "ADD" and "RUN tar" with a RUN command that
uses wget to fetch the file and pipes it into curl. Docker decides to
use the cached layer without fetching the file so it's quite a bit
faster and more bandwidth-efficient.

Signed-off-by: Toby Cabot <toby@acnodal.io>